Move library package to repo root
Import path changes from github.com/ammiranda/otf_api/otf_api
to github.com/ammiranda/otf_api. Users now import as:
  import "github.com/ammiranda/otf_api"
